+++ Client/Components/Pages/About.razor
@@ -9,7 +9,7 @@
     <span class="about-title about-element">IF-Tools Briefings</span>
     <span class="about-version about-element">@(Version ?? "N/A")</span>
     <span class="about-description about-element">A website that allows you to create flight briefings for quick and organized flight planning.</span>
-    <span class="about-copyright about-element">Copyright &copy; 2021 Gleb Zhynko</span>
+    <span class="about-copyright about-element">Copyright &copy; 2024 Gleb Zhynko</span>
     <span class="about-licensed about-element">Licensed under <a href="https://github.com/if-tools/Briefings/blob/master/LICENSE" target="_blank">MIT</a></span>
 </div>
