+++ Client/Components/Pages/ViewBriefing.razor.css
