+++ Client/Components/Shared/BriefingLayout.razor
