+++ Client/Components/Shared/BriefingLayout.razor.css
