+++ Client/Components/Shared/InfoDialog.razor
