+++ Client/Components/_Imports.razor
@@ -9,5 +9,7 @@
 @using IFToolsBriefings.Shared.Data.Models
 @using IFToolsBriefings
 @using IFToolsBriefings.Client.Shared
-@using IFToolsBriefings.Client.Shared.Components
+@using IFToolsBriefings.Client.Components
+@using IFToolsBriefings.Client.Components.Shared
+
