+++ Client/IF-Tools Briefings.Client.csproj
@@ -22,4 +22,32 @@
     <None Remove="Properties\launchSettings.json" />
   </ItemGroup>
 
+  <ItemGroup>
+    <_ContentIncludedByDefault Remove="Pages\About.razor" />
+    <_ContentIncludedByDefault Remove="Pages\EditBriefing.razor" />
+    <_ContentIncludedByDefault Remove="Pages\FindBriefing.razor" />
+    <_ContentIncludedByDefault Remove="Pages\Index.razor" />
+    <_ContentIncludedByDefault Remove="Pages\NewBriefing.razor" />
+    <_ContentIncludedByDefault Remove="Pages\ViewBriefing.razor" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <AdditionalFiles Include="Components\Pages\About.razor" />
+    <AdditionalFiles Include="Components\Pages\EditBriefing.razor" />
+    <AdditionalFiles Include="Components\Pages\FindBriefing.razor" />
+    <AdditionalFiles Include="Components\Pages\Index.razor" />
+    <AdditionalFiles Include="Components\Pages\NewBriefing.razor" />
+    <AdditionalFiles Include="Components\Pages\ViewBriefing.razor" />
+    <AdditionalFiles Include="Components\Shared\AuthenticationModal.razor" />
+    <AdditionalFiles Include="Components\Shared\BriefingLayout.razor" />
+    <AdditionalFiles Include="Components\Shared\GetFlightPlanModal.razor" />
+    <AdditionalFiles Include="Components\Shared\HelpTooltip.razor" />
+    <AdditionalFiles Include="Components\Shared\InfoDialog.razor" />
+    <AdditionalFiles Include="Components\Shared\LoadingIndicator.razor" />
+    <AdditionalFiles Include="Components\Shared\MainLayout.razor" />
+    <AdditionalFiles Include="Components\Shared\NavMenu.razor" />
+    <AdditionalFiles Include="Components\Shared\NotFoundPage.razor" />
+    <AdditionalFiles Include="Components\Shared\TimeSelect.razor" />
+  </ItemGroup>
+
 </Project>
